+++ src/tools/cost-providers/get-cost-provider-accounts.ts
@@ -1,8 +1,8 @@
 import z from "zod";
-import paginationData from "../utils/paginationData";
-import { DEFAULT_LIMIT } from "./structure/constants";
-import MCPUserError from "./structure/MCPUserError";
-import registerTool from "./structure/registerTool";
+import paginationData from "../../utils/paginationData";
+import { DEFAULT_LIMIT } from "../structure/constants";
+import MCPUserError from "../structure/MCPUserError";
+import registerTool from "../structure/registerTool";
 
 const description = `
 Get the list of Cost Provider Accounts that the current user has access to in their workspace and their names.
