+++ src/tools/cost-providers/index.ts
@@ -1,0 +1,2 @@
+import "./get-cost-provider-accounts";
+import "./list-cost-providers";
