+++ src/tools/cost-providers/list-cost-providers.ts
@@ -1,7 +1,7 @@
 import z from "zod";
-import paginationData from "../utils/paginationData";
-import MCPUserError from "./structure/MCPUserError";
-import registerTool from "./structure/registerTool";
+import paginationData from "../../utils/paginationData";
+import MCPUserError from "../structure/MCPUserError";
+import registerTool from "../structure/registerTool";
 
 const description = `
 List of cost providers available to query for a given Workspace. Can be used to filter costs down to a specific provider in VQL queries.
