+++ src/tools/index.ts
@@ -7,6 +7,7 @@ import "./budgets";
 import "./business-metrics";
 import "./canvases";
 import "./cost-alerts";
+import "./cost-providers";
 import "./cost-reports";
 import "./cost-services";
 import "./costs";
@@ -15,9 +16,7 @@ import "./current-user";
 import "./dashboards";
 import "./financial-commitment-reports";
 import "./folders";
-import "./get-cost-provider-accounts";
 import "./list-cost-integrations";
-import "./list-cost-providers";
 import "./network-flow-reports";
 import "./provider-resources";
 import "./recommendation-views";
